third_party/WebKit/LayoutTests/fast/canvas-api/OffscreenCanvas-constructor-in-worker-expected.txt

Tests that the OffscreenCanvas can be constructed on a worker thread.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS width is 50
PASS height is 50
PASS width is 100
PASS height is 100
PASS successfullyParsed is true

TEST COMPLETE
